+++ examples/main/modules.tf
@@ -1,0 +1,141 @@
+module "azure_region" {
+  source  = "claranet/regions/azurerm"
+  version = "x.x.x"
+
+  azure_region = var.azure_region
+}
+
+module "rg" {
+  source  = "claranet/rg/azurerm"
+  version = "x.x.x"
+
+  location    = module.azure_region.location
+  client_name = var.client_name
+  environment = var.environment
+  stack       = var.stack
+}
+
+module "run_common" {
+  source  = "claranet/run-common/azurerm"
+  version = "x.x.x"
+
+  client_name    = var.client_name
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
+  environment    = var.environment
+  stack          = var.stack
+
+  resource_group_name = module.rg.resource_group_name
+
+  tenant_id = var.azure_tenant_id
+
+  monitoring_function_splunk_token = null
+}
+
+resource "azurerm_storage_account" "assets_storage" {
+  account_replication_type = "LRS"
+  account_tier             = "Standard"
+  location                 = module.azure_region.location
+  name                     = "appserviceassets"
+  resource_group_name      = module.rg.resource_group_name
+}
+
+resource "azurerm_storage_share" "assets_share" {
+  name                 = "assets"
+  storage_account_name = azurerm_storage_account.assets_storage.name
+  quota                = 50
+}
+
+module "app_service_plan" {
+  source  = "claranet/app-service-plan/azurerm"
+  version = "x.x.x"
+
+  client_name         = var.client_name
+  environment         = var.environment
+  location            = module.azure_region.location
+  location_short      = module.azure_region.location_short
+  resource_group_name = module.rg.resource_group_name
+  stack               = var.stack
+
+  logs_destinations_ids = [
+    module.run_common.logs_storage_account_id,
+    module.run_common.log_analytics_workspace_id
+  ]
+
+  sku = {
+    tier = "Standard"
+    size = "S1"
+  }
+
+  kind = "Linux"
+}
+
+module "app_service" {
+  source  = "claranet/app-service-web/azurerm"
+  version = "x.x.x"
+
+  client_name         = var.client_name
+  environment         = var.environment
+  location            = module.azure_region.location
+  location_short      = module.azure_region.location_short
+  resource_group_name = module.rg.resource_group_name
+  stack               = var.stack
+
+  app_service_plan_id = module.app_service_plan.app_service_plan_id
+
+  app_settings = {
+    foo = "bar"
+  }
+
+  auth_settings = {
+    enabled             = true
+    token_store_enabled = true
+
+    active_directory = {
+      client_id         = "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
+      client_secret     = "xxxxxxxxxxxxxxxxxxxxx"
+      allowed_audiences = ["https://www.example.com"]
+    }
+  }
+
+  custom_domains = {
+    # Custom domain with SSL certificate file
+    "example.com" = {
+      certificate_file     = "./example.com.pfx"
+      certificate_password = "xxxxxxxxx"
+    }
+    # Custom domain with SSL certificate stored in a keyvault
+    "example.com" = {
+      certificate_keyvault_id = var.certificate_keyvault_id
+    }
+    # Custom domain without SSL certificate
+    "example2.com" = null
+  }
+
+  authorized_ips = ["1.2.3.4/32", "4.3.2.1/32"]
+
+  ip_restriction_headers = {
+    x_forwarded_host = ["myhost1.fr", "myhost2.fr"]
+  }
+  scm_ip_restriction_headers = {
+    x_forwarded_host = ["myhost1.fr", "myhost2.fr"]
+  }
+
+  extra_tags = {
+    foo = "bar"
+  }
+
+  mount_points = [
+    {
+      account_name = azurerm_storage_account.assets_storage.name
+      share_name   = azurerm_storage_share.assets_share.name
+      access_key   = azurerm_storage_account.assets_storage.primary_access_key
+      mount_path   = "/var/www/html/assets"
+    }
+  ]
+
+  logs_destinations_ids = [
+    module.run_common.logs_storage_account_id,
+    module.run_common.log_analytics_workspace_id
+  ]
+}